% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cDC1}
\alias{cDC1}
\title{cDC1 adjacency matrix}
\format{An 8568 x 8568 data matrix containing pairwise confidence score for each protein pair}
\usage{
data(cDC1_adj_matrix)
}
\description{
An adjacency matrix of the proteins in the cDC1 cell containing the confidence
scores of each pairiwse protein-protein interaction (PPI). The confidence scores
have been obtained from PPI interactome InWeb_IM.
}
\examples{
data(cDC1_adj_matrix)
}
\keyword{adjacency}
