% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowtie_functions.R
\name{consecutive_str}
\alias{consecutive_str}
\title{Find sub-vectors of a certain minimum size and average value in a vector}
\usage{
consecutive_str(input_vector, window_length, min_value)
}
\arguments{
\item{input_vector}{A numeric vector containing values between 0 and 1.}

\item{window_length}{The minimum size of the desired sub-vectors.}

\item{min_value}{The minimum average value of the sub-vectors.}
}
\value{
Returns a list of indicies that specify the location of found sub-vectors in the input vector
}
\description{
Using a sliding window approach, this function returns  sub-vectors of a minimum size and average value found in a numeric vector containing values between 0-1
}
\examples{

}
\keyword{sub-vectors}
