% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowtie_functions.R
\name{cut_into_chunks}
\alias{cut_into_chunks}
\title{Cuts a vector of indicies into sub-vectors that are consecutive (increase by +1)}
\usage{
cut_into_chunks(string_vect)
}
\arguments{
\item{string_vect}{A vector of indicies}
}
\value{
Returns a list of vectors containing indicies that are consecutive
}
\description{
Cuts a vector of indicies into sub-vectors that are consecutive (increase by +1)
}
\examples{

}
\keyword{submatrices}
