% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowtie_functions.R
\name{find_bowties}
\alias{find_bowties}
\title{Find strings of high-confidence interactions between protein complexes in a first-order direct interaction matrix}
\usage{
find_bowties(
  full_matrix,
  protein_complex_areas,
  int_fan_min_size = 10,
  int_fan_avg_conf_score = 0.9
)
}
\arguments{
\item{full_matrix}{A first-order direct interaction matrix containing scores between 0-1, where 1 indicates an interaction with highest confidence and 0 indicates no interaction.}

\item{protein_complex_areas}{Start and end ranges along the matrix diagonal annotated to be protein complex submatrices.}

\item{int_fan_min_size}{Minimum number of interactions in a bow-tie interaction fan (default: 10).}

\item{int_fan_avg_conf_score}{Minimum average confidence score across the interactions in a bow-tie interaction fan (default: 0.9).}
}
\value{
Returns a list of "knot" proteins and their interactions fans indicated by start and end range indicies.
}
\description{
Find strings of high-confidence interactions between protein complexes in a first-order direct interaction matrix.
}
\examples{

}
\keyword{submatrices}
