% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowtie_functions.R
\name{find_complexes}
\alias{find_complexes}
\title{Locate submatrices (protein complexes) on the diagonal of a first-order direct interaction matrix}
\usage{
find_complexes(
  full_matrix,
  min_conf_score = 0.75,
  max_complex_size = 500,
  give_up_expand = 50
)
}
\arguments{
\item{full_matrix}{A first-order direct interaction matrix containing scores between 0-1, where 1 indicates an interaction with highest confidence and 0 indicates no interaction.}

\item{min_conf_score}{The minimum average score of a submatrix to be accepted as a protein complex (default: 0.75).}

\item{max_complex_size}{The maximum size of a submatrix (default: 500).}

\item{give_up_expand}{Give up expanding a submatrix for a given seed when exceeding N proteins in each direction (default: 50). This parameter was implemented to reduce computing time.}
}
\description{
Locate high-scoring square submatrices along the diagonal of a first-order direct interaction matrix.
}
\examples{

}
\keyword{submatrices}
