% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowtie_functions.R
\name{from_list_to_df}
\alias{from_list_to_df}
\title{Tidy-up a list of wTO scores, as given by wTO.network}
\usage{
from_list_to_df(score_list)
}
\arguments{
\item{score_list}{A list containing three vectors: The names of node1, names of node2 and pairwise wTO scores}
}
\description{
Tidy up a list of wTO scores, returning a complete pairwise matrix
}
\examples{

# Create graph
random_graph           = sample_pa(n=200, power = 1.2, directed=F)
E(random_graph)$weight = runif(n=length(E(random_graph)))

# Calculate wTO
wTO_list   = bowtie::wTO.network(node_vector = as.vector(V(random_graph)), igraph_object = random_graph, thread_numb = 4)
wTO_matrix = bowtie::from.list.to.df(wTO_list)

# Ordering of nodes
hclust_object = hclust(as.dist(1-wTO_matrix), method = 'average')
node_order    = hclust_object$labels[hclust_object$order]

# Visualize
image(wTO_matrix[node_order,node_order], useRaster = T, col = colorRampPalette(brewer.pal(9,"YlGnBu"))(49))

}
\keyword{overlap,}
\keyword{pairwise}
\keyword{topological}
\keyword{wTO,}
\keyword{weighted}
